add logger

+++ esbox/models/torch_model.py
@@ -57,7 +57,7 @@ class TorchModel(nn.Module):
         i = 0
         arrays = []
         for shape in array_shapes:
-            size = np.prod(shape, dtype=np.int)
+            size = np.prod(shape, dtype=np.int32)
             array = flat_array[i:(i + size)].reshape(shape)
             arrays.append(array)
             i += size
